refactor/list: enter, tab 1차 리팩토링
