+++ src/frontend/src/component/editor/context/NoteDataProvider.jsx
@@ -20,7 +20,7 @@ export const BlockIdListContext = createContext(null);
 function NoteDataProvider({children, noteId}) {
     const [blockIdList, setBlockIdList] = useState([]);
     const blockStore = useRef(null);
-    console.log(blockStore);
+    window.blockStore = blockStore.current;
     const navigate = useNavigate();
 
     useEffect(() => {
