+++ src/frontend/src/component/editor/text/hooks/useTextHandler.jsx
@@ -1,8 +1,40 @@
 import {editorSelection} from "../../../../App";
 import {useBlockStore} from "../../hooks/useBlockHooks";
+import {Text} from "../../../../model/Text";
+import {getRandomId} from "../../../../utils/id";
+import {FontStyle} from "../../../../model/FontStyle";
 
 function useTextHandler() {
     const blockStore = useBlockStore();
+
+    /**
+     * Text 를 분리하고 분리된 텍스트 객체들을 반환
+     */
+    const divideText = (block) => {
+        // 분리하고 업데이트된 textIdx 구함
+        const text = block.getTextFromId(editorSelection.getClosestId("text").start);
+        let textIdx = block.getTextIdx(text.id);
+        const dividedTextContents = editorSelection.getDividedTextContentsFromCaret();
+        const cnt = block.divideText(textIdx, dividedTextContents.before, dividedTextContents.after);
+
+        // text 마지막에 캐럿이 잡힐 경우 예외 처리
+        if (dividedTextContents.after === "") textIdx++;
+
+        // 추가된만큼 idx 증가
+        textIdx += cnt;
+
+        // 기존 textBlock에 있는 text들 삭제
+        const removedTextList = [];
+        while (true) {
+            const text = block.removeText(textIdx);
+            if (!text) break;
+            removedTextList.push(text);
+        }
+        if (block.textIdList.length === 0) {
+            block.addText(new Text(getRandomId(), "", new FontStyle()));
+        }
+        return removedTextList;
+    }
 
     const updateTextValue = () => {
         // 실제돔 불러오기
@@ -39,7 +71,7 @@ function useTextHandler() {
         }
     }
 
-    return {updateTextValue, deleteTextValue}
+    return {updateTextValue, deleteTextValue, divideText}
 }
 
 export default useTextHandler;
